# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TradeUpdate API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/chart/candles:
    get:
      operationId: getCandles
      tags: [user]
      summary: Get candlestick data
      parameters:
        - name: timeframe
          in: query
          required: true
          schema:
            type: string
            enum: ["1m", "5m", "15m"]
        - name: count
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Candle data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandleListResponse"
...
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: List all users
      parameters:
        - name: filter
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
...
  /admin/users/{userId}:
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Delete user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/users/{userId}/dashboard:
    get:
      operationId: adminGetUserDashboard
      tags: [admin]
      summary: View user dashboard as admin
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User dashboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDashboard"
...
  /admin/strategies/{strategyId}:
    patch:
      operationId: updateStrategy
      tags: [admin]
      summary: Update strategy
      parameters:
        - name: strategyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
    delete:
      operationId: deleteStrategy
      tags: [admin]
      summary: Delete strategy
      parameters:
        - name: strategyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/strategies/{strategyId}/pause:
    post:
      operationId: pauseStrategy
      tags: [admin]
      summary: Pause strategy
      parameters:
        - name: strategyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/strategies/{strategyId}/reactivate:
    post:
      operationId: reactivateStrategy
      tags: [admin]
      summary: Reactivate strategy
      parameters:
        - name: strategyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /admin/strategies/{strategyId}/performance:
    get:
      operationId: getStrategyPerformance
      tags: [admin]
      summary: Get strategy performance data
      parameters:
        - name: strategyId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Performance data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StrategyPerformanceData"
  /admin/copy-trade/preview:
    post:
      operationId: previewCopyTrade
      tags: [admin]
      summary: Preview copy trade before executing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CopyTradeInput"
      responses:
        "200":
          description: Copy trade preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CopyTradePreview"
  /admin/copy-trade/execute:
    post:
      operationId: executeCopyTrade
      tags: [admin]
      summary: Execute copy trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CopyTradeInput"
      responses:
        "200":
          description: Copy trade executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /admin/backtest/run:
    post:
      operationId: runBacktest
      tags: [admin]
      summary: Run a backtest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BacktestInput"
      responses:
        "200":
          description: Backtest results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestResult"
...
  /admin/settings/master-stop:
    post:
      operationId: setMasterStop
      tags: [admin]
      summary: Toggle master stop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MasterStopInput"
      responses:
        "200":
          description: Master stop updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
        success:
          type: boolean
...
    CurrentUser:
      type: object
      required: [id, username, role, status]
      properties:
        id:
          type: string
        username:
          type: string
        role:
          type: string
        status:
          type: string
        email:
          type: string
        tradingProfile:
          type: ["string", "null"]
        strategyId:
          type: ["string", "null"]
        accountBalance:
          type: number
        peakBalance:
          type: number
        dailyStartBalance:
          type: number
        autoCompoundEnabled:
          type: boolean
        adaptiveIntelligenceEnabled:
          type: boolean
        copyTradingEnabled:
          type: boolean
        mustChangePassword:
          type: boolean
        hasDerivToken:
          type: boolean
        tradingMode:
          type: ["string", "null"]
...
    UserDashboard:
      type: object
      required: [user, botStatus, recentTrades, sessionPerformance]
      properties:
        user:
          $ref: "#/components/schemas/CurrentUser"
        botStatus:
          $ref: "#/components/schemas/BotStatus"
        recentTrades:
          type: array
          items:
            $ref: "#/components/schemas/Trade"
        sessionPerformance:
          type: array
          items:
            $ref: "#/components/schemas/SessionSummary"
        dailyPnl:
          type: number
        todayTrades:
          type: integer
        winStreak:
          type: integer
        lossStreak:
          type: integer
...
    SessionSummary:
      type: object
      required: [sessionName, totalTrades, wins, losses, winRate, sessionMultiplier]
      properties:
        sessionName:
          type: string
        totalTrades:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        winRate:
          type: number
        sessionMultiplier:
          type: number
...
    IndicatorAccuracy:
      type: object
      properties:
        rsi:
          type: number
        stoch:
          type: number
        macd:
          type: number
        pullback:
          type: number
        smc:
          type: number
    Candle:
      type: object
      required: [time, open, high, low, close, volume]
      properties:
        time:
          type: integer
        open:
          type: number
        high:
          type: number
        low:
          type: number
        close:
          type: number
        volume:
          type: number
    CandleListResponse:
      type: object
      required: [candles, timeframe]
      properties:
        candles:
          type: array
          items:
            $ref: "#/components/schemas/Candle"
        timeframe:
          type: string
...
    AdminUser:
      type: object
      required: [id, username, email, role, status, isActive]
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        role:
          type: string
        status:
          type: string
        isActive:
          type: boolean
        tradingProfile:
          type: ["string", "null"]
        strategyId:
          type: ["string", "null"]
        strategyName:
          type: ["string", "null"]
        accountBalance:
          type: number
        peakBalance:
          type: number
        dailyStartBalance:
          type: number
        copyTradingEnabled:
          type: boolean
        autoCompoundEnabled:
          type: boolean
        adaptiveIntelligenceEnabled:
          type: boolean
        country:
          type: ["string", "null"]
        hasDerivToken:
          type: boolean
        tradingMode:
          type: ["string", "null"]
        botRunning:
          type: boolean
        recoveryModeActive:
          type: boolean
        winStreakCautionActive:
          type: boolean
        createdAt:
          type: ["integer", "null"]
        lastLogin:
          type: ["integer", "null"]
        approvedAt:
          type: ["integer", "null"]
        approvedBy:
          type: ["string", "null"]
    AdminUserListResponse:
      type: object
      required: [users]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
...
    AdminUserUpdate:
      type: object
      properties:
        status:
          type: string
        isActive:
          type: boolean
        tradingProfile:
          type: string
        strategyId:
          type: string
        copyTradingEnabled:
          type: boolean
        accountBalance:
          type: number
...
    StrategyUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
        scoreThreshold:
          type: integer
        maxTradesDay:
          type: integer
        maxRiskPercent:
          type: number
        autoPauseOnLosses:
          type: integer
    StrategyPerformanceData:
      type: object
      required: [strategy, dailyPerformance, sessionBreakdown, indicatorAccuracy, assignedUsers]
      properties:
        strategy:
          $ref: "#/components/schemas/Strategy"
        dailyPerformance:
          type: array
          items:
            $ref: "#/components/schemas/DailyPerformance"
        sessionBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/SessionSummary"
        indicatorAccuracy:
          $ref: "#/components/schemas/IndicatorAccuracy"
        assignedUsers:
          type: array
          items:
            $ref: "#/components/schemas/StrategyUserSummary"
    DailyPerformance:
      type: object
      required: [date, trades, wins, losses, pnl, avgScore]
      properties:
        date:
          type: integer
        trades:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        pnl:
          type: number
        avgScore:
          type: number
    StrategyUserSummary:
      type: object
      required: [userId, username, winRate, totalTrades]
      properties:
        userId:
          type: string
        username:
          type: string
        winRate:
          type: number
        totalTrades:
          type: integer
    CopyTradeInput:
      type: object
      required: [direction, targetUserIds, riskMultiplier]
      properties:
        direction:
          type: string
          enum: [BUY, SELL]
        targetUserIds:
          type: array
          items:
            type: string
        riskMultiplier:
          type: number
        forceOverrideUserIds:
          type: array
          items:
            type: string
    CopyTradePreview:
      type: object
      required: [willReceive, willSkip, forceAvailable, summary]
      properties:
        willReceive:
          type: array
          items:
            $ref: "#/components/schemas/CopyTradeUserPreview"
        willSkip:
          type: array
          items:
            $ref: "#/components/schemas/CopyTradeSkipPreview"
        forceAvailable:
          type: array
          items:
            $ref: "#/components/schemas/CopyTradeUserPreview"
        summary:
          $ref: "#/components/schemas/CopyTradeSummary"
...
    BacktestInput:
      type: object
      required: [strategyId, dateFrom, dateTo]
      properties:
        strategyId:
          type: string
        dateFrom:
          type: integer
        dateTo:
          type: integer
    BacktestResult:
      type: object
      required: [id, strategyId, totalTrades, wins, losses, winRate, profitFactor, maxDrawdown, totalPnl]
      properties:
        id:
          type: string
        strategyId:
          type: string
        runBy:
          type: string
        dateFrom:
          type: integer
        dateTo:
          type: integer
        totalTrades:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        winRate:
          type: number
        profitFactor:
          type: number
        maxDrawdown:
          type: number
        totalPnl:
          type: number
        equityCurve:
          type: array
          items:
            $ref: "#/components/schemas/EquityPoint"
        sessionBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/SessionSummary"
        indicatorAccuracy:
          $ref: "#/components/schemas/IndicatorAccuracy"
        scoreRangePerformance:
          type: array
          items:
            $ref: "#/components/schemas/ScoreRangeSummary"
        bestTrade:
          type: number
        worstTrade:
          type: number
        avgDurationMinutes:
          type: number
        sharpeRatio:
          type: number
        createdAt:
          type: integer
...
    MasterStopInput:
      type: object
      required: [active]
      properties:
        active:
          type: boolean
...